backend/requirements.txt
absl-py==0.15.0
amqp==5.0.9
asgiref==3.5.0
astor 
astunparse==1.6.3
billiard==3.6.4.0
cached-property 
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.12
django-celery-results==2.3.0
django-cors-headers==3.11.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
dnspython==2.2.0
eventlet==0.33.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio 
h5py==3.1.0
idna==3.3
importlib-metadata
Keras-Applications==1.0.8
Keras-Preprocessing 
kombu==5.2.3
Markdown 
mysqlclient==2.1.0
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy 
six==1.15.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-estimator==2.8.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
Werkzeug 
wrapt==1.12.1
zipp
django_redis
django-redis-cache 
